% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Y-data.R
\docType{data}
\name{Y}
\alias{Y}
\title{A dataset containing anomaly scores}
\format{
A dataframe of 803 rows and 9 columns
\describe{
 \item{Dimension 1}{The different anomaly detection methods}
 \item{Dimension 2}{Each column has the anomaly scores from that method for the annulus dataset.}
}
}
\usage{
Y
}
\description{
This dataset contains anomaly scores for the annulus data
}
\keyword{datasets}
