% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yfaithful-data.R
\docType{data}
\name{Yfaithful}
\alias{Yfaithful}
\title{Old Faithful anomaly scores}
\format{
A dataframe of 272 rows and 9 columns
\describe{
 \item{Dimension 1}{The different anomaly detection methods}
 \item{Dimension 2}{Each column has the anomaly scores from that method for the old faithful dataset.}
}
}
\usage{
Yfaithful
}
\description{
This dataset contains anomaly scores for the Old Faithful dataset
}
\keyword{datasets}
